requirements.txt
pytest == 9.0.2
allure-pytest
selene == 2.0.0rc9
selenium == 4.39.0
